feat: added authorization options for admin and workers
