+++ src/Controller/ProductController.cs
@@ -7,12 +7,14 @@ using ControlDeVenta_Proy.src.Helpers;
 using ControlDeVenta_Proy.src.Interfaces;
 using ControlDeVenta_Proy.src.Mappers;
 using ControlDeVenta_Proy.src.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ControlDeVenta_Proy.src.Controller
 {
     [ApiController]
     [Route("api/[controller]")]
+    [Authorize(Roles = "Admin, Worker")]
     public class ProductController : ControllerBase
     {
         private readonly IProductRepository _productRepository;
@@ -29,48 +31,6 @@ namespace ControlDeVenta_Proy.src.Controller
             {
                 var products = await _productRepository.GetProducts(query);
                 return Ok(products);
-            }
-            catch (Exception e)
-            {
-                return BadRequest(e.Message);
-            }
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> AddProduct([FromBody] NewPorductDto product)
-        {
-            try
-            {
-                var newProduct = await _productRepository.AddProduct(product.MapToCreateProduct());
-                return Ok(newProduct);
-            }
-            catch (Exception e)
-            {
-                return BadRequest(e.Message);
-            }
-        }
-
-        [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateProduct(int id, [FromBody] NewPorductDto product)
-        {
-            try
-            {
-                var updatedProduct = await _productRepository.UpdateProduct(id, product);
-                return Ok(updatedProduct);
-            }
-            catch (Exception e)
-            {
-                return BadRequest(e.Message);
-            }
-        }
-
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteProduct(int id)
-        {
-            try
-            {
-                var deletedProduct = await _productRepository.DeleteProduct(id);
-                return Ok(deletedProduct);
             }
             catch (Exception e)
             {
